Support SRV and DKIM records (#1)
* Allow underscore characters in FQDNs and pointers.

Modern DNS records can contain underscores for a number of reasons. In
particular, DKIM records involve a "_domainkey" part,

  https://tools.ietf.org/html/rfc6376

that is rejected by the current "fqdn" and "p" validation routines.
Moreover, any SRV records will have a service name prefixed with an
underscore:

  https://tools.ietf.org/html/rfc2782

To recognize these tokens as valid, this commit expands the "fqdn" and
"p" regular expressions to allow underscores as the first character in
each component of an FQDN.

* Add support for SRV records.

There is a patch for djbdns that adds support for SRV records to both
tinydns-data and axfr-get:

  From: Michael Handler <handler@sub-rosa.com>
  To: dns@list.cr.yp.to
  Subject: tinydns-data SRV & axfr-get SRV/PTR patches
  Date: Thu, 14 Sep 2000 20:37:50 -040

Many distributions carry the patch, but valtz rejects the SRV records
because it doesn't recognize the "S" indicator or know how to validate
the port, weight, or priority fields.

This commit adds support for the new record type, and adds validation
routines for the three new fields. All of them are the same: ports,
weights, and priorities are all integers between 0 and 65536.

* Whitespace cleanup.

Remove some trailing whitespace and empty comments where it should (ha
ha) have no effect on the behavior of the program.